Refactor  `UIWebView` extension to `FolioReaderWebView`
